fix type errors and circular dependencies
