+++ snuba/query/validation/signature.py
@@ -16,7 +16,7 @@ from snuba.clickhouse.columns import (
     String,
     UInt,
 )
-from snuba.query.data_source import ColumnSet, DataSource
+from snuba.query.data_source import DataSource
 from snuba.query.expressions import Expression
 from snuba.query.expressions import Literal as LiteralType
 from snuba.query.matchers import Any as AnyMatcher
@@ -24,6 +24,7 @@ from snuba.query.matchers import Column as ColumnMatcher
 from snuba.query.matchers import Literal as LiteralMatcher
 from snuba.query.matchers import Param
 from snuba.query.validation import FunctionCallValidator, InvalidFunctionCall
+from snuba.utils.schemas import ColumnSet
 
 logger = logging.getLogger(__name__)
